encapsulate the coinductive logic into chalk-solve
This also removes the need to work around the lack of trait object
upcasts.
